+++ front/src/modules/companies/hooks/useOpenActionBar.tsx
@@ -5,9 +5,9 @@ import { useOpenCreateActivityDrawerForSelectedRowIds } from '@/activities/hooks
 import { ActivityTargetableEntityType } from '@/activities/types/ActivityTargetableEntity';
 import { GET_PIPELINES } from '@/pipeline/queries';
 import { ActionBarEntry } from '@/ui/action-bar/components/ActionBarEntry';
+import { actionBarEntriesState } from '@/ui/action-bar/states/ActionBarEntriesState';
 import { IconCheckbox, IconNotes, IconTrash } from '@/ui/icon';
 import { useResetTableRowSelection } from '@/ui/table/hooks/useResetTableRowSelection';
-import { actionBarEntriesState } from '@/ui/table/states/ActionBarEntriesState';
 import { selectedRowIdsSelector } from '@/ui/table/states/selectedRowIdsSelector';
 import { tableRowIdsState } from '@/ui/table/states/tableRowIdsState';
 import {
