+++ front/src/modules/people/hooks/useOpenActionBar.tsx
@@ -4,9 +4,9 @@ import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
 import { useOpenCreateActivityDrawerForSelectedRowIds } from '@/activities/hooks/useOpenCreateActivityDrawerForSelectedRowIds';
 import { ActivityTargetableEntityType } from '@/activities/types/ActivityTargetableEntity';
 import { ActionBarEntry } from '@/ui/action-bar/components/ActionBarEntry';
+import { actionBarEntriesState } from '@/ui/action-bar/states/ActionBarEntriesState';
 import { IconCheckbox, IconNotes, IconTrash } from '@/ui/icon';
 import { useResetTableRowSelection } from '@/ui/table/hooks/useResetTableRowSelection';
-import { actionBarEntriesState } from '@/ui/table/states/ActionBarEntriesState';
 import { selectedRowIdsSelector } from '@/ui/table/states/selectedRowIdsSelector';
 import { tableRowIdsState } from '@/ui/table/states/tableRowIdsState';
 import { ActivityType, useDeleteManyPersonMutation } from '~/generated/graphql';
