+++ front/src/modules/ui/action-bar/states/ActionBarEntriesState.ts
