+++ front/src/modules/ui/action-bar/states/ActionBarIsOpenState.ts
