+++ front/src/modules/ui/context-menu/components/__stories__/ActionBar.stories.tsx
@@ -1,0 +1,17 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import { ComponentDecorator } from '~/testing/decorators/ComponentDecorator';
+
+import { ContextMenu } from '../ContextMenu';
+
+const meta: Meta<typeof ContextMenu> = {
+  title: 'UI/ContextMenu/ContextMenu',
+  component: ContextMenu,
+  decorators: [ComponentDecorator],
+  args: { selectedIds: [] },
+};
+
+export default meta;
+type Story = StoryObj<typeof ContextMenu>;
+
+export const Default: Story = {};
