+++ front/src/modules/ui/context-menu/states/ContextMenuEntriesState.ts
