+++ front/src/modules/ui/context-menu/states/ContextMenuIsOpenState.ts
